// types/enum/store/wallet/mutations.enum.ts
export enum WalletMutationsEnum {
  SET_APP_MODE = 'SET_APP_MODE',
  SET_WALLET_STATE = 'SET_WALLET_STATE',
  SET_CONTRACT_ADDRESS = 'SET_CONTRACT_ADDRESS',
  SET_CONTRACT_ADDRESS_NEW = 'SET_CONTRACT_ADDRESS_NEW',
  SET_CONTRACT_INFO = 'SET_CONTRACT_INFO',
  SET_WEB3 = 'SET_WEB3',
  SET_WEB3_MODAL = 'SET_WEB3_MODAL',
  SET_NETWORK_ID = 'SET_NETWORK_ID',
  SET_WALLET_DATA = 'SET_WALLET_DATA',
  SET_REF_WALLET_DATA = 'SET_REF_WALLET_DATA',
  SET_PROVIDER_NAME = 'SET_PROVIDER_NAME',
  SET_METAMASK_INSTALLATION_LINK = 'SET_METAMASK_INSTALLATION_LINK',
  SET_BINANCE_SMART_CHAIN_INSTALLATION_LINK = 'SET_BINANCE_SMART_CHAIN_INSTALLATION_LINK',
  RESET_WALLET_STATE = 'RESET_WALLET_STATE',
}