// types/enum/store/wallet/actions.enum.ts
export enum WalletActionsEnum {
  initWeb3Modal = 'initWeb3Modal',
  initConnect = 'initConnect',
  connectWallet = 'connectWallet',
  disconnectWallet = 'disconnectWallet',
  metamaskConnect = 'metamaskConnect',
  bscConnect = 'bscConnect',
  walletConnect = 'walletConnect',
  updateAccount = 'updateAccount',
  subscribeProvider = 'subscribeProvider',
  setRefWalletAddress = 'setRefWalletAddress',
  setNetworkId = 'setNetworkId',
  checkWallets = 'checkWallets',
  checkBalance = 'checkBalance',
  contractDays = 'contractDays',
  investors = 'investors',
  totalInvested = 'totalInvested',
  getUserLevel = 'getUserLevel',
}