// types/enum/store/wallet/getters.enum.ts
export enum WalletGettersEnum {
  appMode = 'wallet/appMode',
  defaultProvider = 'wallet/defaultProvider',
  provider = 'wallet/provider',
  ballance = 'wallet/ballance',
  contracts = 'wallet/contracts',
  contractInstance = 'wallet/contractInstance',
  contractInstanceNew = 'wallet/contractInstanceNew',
  contractAddress = 'wallet/contractAddress',
  contractAddressNew = 'wallet/contractAddressNew',
  contractOwnerAddress = 'wallet/contractOwnerAddress',
  contracInfo = 'wallet/contracInfo',
  web3 = 'wallet/web3',
  walletAddress = 'wallet/walletAddress',
  walletAddressSlice = 'wallet/walletAddressSlice',
  refWalletAddress = 'wallet/refWalletAddress',
  metamaskInstallationLink = 'wallet/metamaskInstallationLink',
  binanceSmartChainInstallationLink = 'wallet/binanceSmartChainInstallationLink',
  isMetamaskInstalled = 'wallet/isMetamaskInstalled',
  isBinanceSmartChainInstalled = 'wallet/isBinanceSmartChainInstalled',
  isMetamaskAvailable = 'wallet/isMetamaskAvailable',
  isBinanceSmartChainWalletAvailable = 'wallet/isBinanceSmartChainWalletAvailable',
  isWalletConnectAvailable = 'wallet/isWalletConnectAvailable',
}